Extracted clientError & various form helpers into own files
