+++ management-interface.go
@@ -700,53 +700,3 @@ func CameraHandler(w http.ResponseWriter, r *http.Request) {
 func CameraSnapshot(w http.ResponseWriter, r *http.Request) {
 	http.ServeFile(w, r, "/var/spool/cptv/still.png")
 }
-
-func newClientError(msg string) *clientError {
-	return &clientError{msg}
-}
-
-type clientError struct {
-	msg string
-}
-
-func (e *clientError) Error() string {
-	return e.msg
-}
-
-func isClientError(err error) bool {
-	_, ok := err.(*clientError)
-	return ok
-}
-
-func trimmedFormValue(r *http.Request, name string) string {
-	return strings.TrimSpace(r.FormValue(name))
-}
-
-func parseFloat(val string) (float64, bool) {
-	f, err := strconv.ParseFloat(val, 64)
-	if err != nil {
-		return 0, false
-	}
-	return f, true
-}
-
-func floatToString(val float64) string {
-	if val == 0 {
-		return ""
-	}
-	return fmt.Sprint(val)
-}
-
-func successMessage(err error, msg string) string {
-	if err == nil {
-		return msg
-	}
-	return ""
-}
-
-func errorMessage(err error) string {
-	if err == nil {
-		return ""
-	}
-	return err.Error()
-}
